Improving code for the new subnets dropdown
